node-main/test/fixtures/wpt/url/a-element.html

<!doctype html>
<meta charset=utf-8>
<meta name="variant" content="?include=file">
<meta name="variant" content="?include=javascript">
<meta name="variant" content="?include=mailto">
<meta name="variant" content="?exclude=(file|javascript|mailto)">
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src="/common/subset-tests-by-key.js"></script>
<base id=base>
<div id=log></div>
<script src=resources/a-element.js></script>
<!--
  Other dependencies:
  * resources/urltestdata.json,
  * resources/urltestdata-javascript-only.json,
-->


<a id="multline-entity" download="multline-entity.txt" href="data:text/plain;charset=utf-8,first%20line&#10;second%20line"> Link with embedded \n is parsed correctly </a>

<script type="text/javascript">
test(function() {
  const link = document.getElementById("multline-entity");
  assert_equals(link.href, "data:text/plain;charset=utf-8,first%20linesecond%20line");
}, "Test that embedded 0x0A is stripped");
</script>
